requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==1.0.2
pandas==2.1.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
